+++ backend/app/views/admin.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, jsonify, session
 from app import db
 from app.controller import scheduler
-from app.utils import check_csrf_token
+from app.tool import check_csrf_token
 
 from app.datamodel import Device, Status  # 获取关于设备的信息
