+++ backend/app/views/client.py
@@ -2,7 +2,7 @@ import time
 from flask import Blueprint, request, jsonify
 from app.datamodel import Device
 from app.controller import client_remote_map, scheduler
-from app.utils import verify_signature
+from app.tool import verify_signature
 
 client_blueprint = Blueprint("client", __name__)
